requirements.txt
aiogram==2.14.3
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
Babel==2.9.1
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.4
idna==3.2
multidict==5.1.0
mysql-connector-python==8.0.26
protobuf==3.17.3
pytz==2021.1
requests==2.26.0
selenium==3.141.0
six==1.16.0
spotipy==2.19.0
typing-extensions==3.10.0.2
urllib3==1.26.6
yarl==1.6.3
